+++ module/chat/chat-processors.js
@@ -40,6 +40,7 @@ export default function RegisterChatProcessors() {
   ChatProcessors.registerProcessor(new SetEventFlagsChatProcessor())
   ChatProcessors.registerProcessor(new RemoteChatProcessor())
   ChatProcessors.registerProcessor(new SlamChatProcessor())
+  //ChatProcessors.registerProcessor(new JB2AChatProcessor())
 }
 
 class SetEventFlagsChatProcessor extends ChatProcessor {
@@ -87,13 +88,15 @@ class RollAgainstChatProcessor extends ChatProcessor {
     return '/ra N | Skillname-N'
   }
   matches(line) {
-    this.match = line.match(/^([\.\/]p?ra) +(\w+-)?(\d+)/i)
+    this.match = line.match(/^([\.\/]p?ra) +([\w-'" ]+-)?(\d+)/i)
     return !!this.match
   }
   async process(line) {
     let m = this.match
     let skill = m[2] || 'Default='
-    let action = parselink('S:' + skill.replace('-', '=') + m[3])
+    skill = skill.replace('-', '=') + m[3]
+    if (skill.includes(' ')) skill = '"' + skill + '"'
+    let action = parselink('S:' + skill)
     this.send() // send what we have
     await GURPS.performAction(action.action, GURPS.LastActor, {
       shiftKey: line.substr(1).startsWith('pra') || this.msgs().event?.shiftKey,
@@ -504,5 +507,39 @@ class TrackerChatProcessor extends ChatProcessor {
         this.prnt(`${i18n('GURPS.chatResourceTracker')}${display} ${m[5]} = ${v}`)
       } else ui.notifications.warn(`${i18n('GURPS.chatUnrecognizedFormat', 'Unrecognized format')} '${line}'`)
     }
+  }
+}
+
+class JB2AChatProcessor extends ChatProcessor {
+  help() {
+    return '/jb2a &lt;animation macro name&gt;'
+  }
+  matches(line) {
+    this.match = line.match(/^\/jb2a +(\d*) *(.*)/i)
+    return !!this.match
+  }
+  async process(line) {
+    let otigons = game.packs.entries.filter(c => c.metadata.module?.startsWith("otigons"))
+    if (otigons.length == 0) {
+      ui.notifications.warn('You must have https://github.com/otigon/otigons-animation-macros loaded')
+      return
+    }
+    const wait = (delay) => new Promise((resolve) => setTimeout(resolve, delay))
+    let t = !!this.match[1] ? parseInt(this.match[1]) : 0
+    let pat = new RegExp(makeRegexPatternFrom(this.match[2], false), 'i')
+    for (const c of otigons) {
+      let found = c.index.find(e => {console.log(e.name); return e.name.match(pat)})
+      if (found) {
+        c.getEntity(found._id).then(async m => {
+          if (!!m) {
+            await wait(t)
+            m.execute()
+          } else
+            ui.notifications.warn(`Entry found for '${this.match[2]}', but no macro`)
+        })
+        return
+      }
+    }
+    ui.notifications.warn(`No macro found for '${this.match[2]}'`)
   }
 }